Fix the test to work on PyPy
It appears that PyPy occasions calls x.__hash__() multiple times during
the execution of hash(x). Rather than making assertions about the
number of times hash is called, introduce a new class that can be put
into a state where it throws an exception when hash() is called. Use
this mechanism to verify hash is never called on the object after the
first hash of the pmap that contains the object.