Add ignore_extra kwarg on PClass, PRecord create class method
This is an optional argument which defaults to False; when set to True,
it allows a dict to be passed in to the create method which contains
keys that are not present in the fields of the PClass or PRecord - these
extraneous keys are simply ignored.

This would be useful for creating an instance of a PClass or PRecord
which only needs to know about a subset of fields from a data source,
possibly one whose format is out of your control.

There are unit tests for both the PClass and PRecord in the cases where
this argument is True and where it is False.

Resolves #128