Only require pytest-runner if running tests
Having pytest-runner in setup_requires forces us to always
install this to run setup.py. This is problematic for two reasons:

1) It's only needed for running the tests, so we don't really
want to install this for non-development cases.
2) If installing the package through pip, setup_requires is handed
off to easy_install - however, this means it then ignores any
pip configuration (so is problematic for using custom indexes
or networking configuration).