Update doc build requirements; test locally
The documentation can now be built and tested using tox,
with warnings escalated to errors.

In addition, a new `update-docs` tox environment exists
which can update the docs build requirements.

The `Makefile` and `make.bat` files are removed,
along with the `docs/.gitignore` file,
because the build is now standardized via tox
and the HTML docs are written to `build/docs`.

Finally, the Read the Docs configuration is updated.

----

The documentation cannot be built in nitpicky mode (`-n`);
this is because unbound `collections.abc` classes' methods
are assigned directly to various class methods,
which allows Sphinx to see -- and fail to parse/render --
the docstrings from the ABC class methods.

For example, the `PDeque` class contains this line:

```
index = Sequence.index
```

However, the `Sequence.index` docstring cannot be rendered
and results in a failure.

It may be possible to enable nitpicky mode at a later date
if changes are made to the source code.